Migrate from Compose to XML

+++ app/src/main/java/com/fadel/go4lunch/ui/list/RestaurantsItemUiModel.kt
@@ -7,6 +7,6 @@ data class RestaurantsItemUiModel(
     val openingHours: String,
     val distance: String,
     val interestNumber: String,
-    val numberOfStars: Double,
+    val numberOfStars: Float,
     val onItemClicked: (() -> Unit)?
 )
